refactor: 테스트 파일 패키지 이동
* 리팩토링 하면서 산재된 테스트 파일들을 제대로 된 위치로 이동한다.
