+++ src/test/java/webserver/http/request/HttpRequestTest.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.http.request;
 
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
