+++ src/test/java/webserver/http/request/RequestBodyTest.java
@@ -1,4 +1,4 @@
-package webserver.http;
+package webserver.http.request;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
