+++ src/test/java/webserver/http/response/ResponseLineTest.java
@@ -1,4 +1,4 @@
-package model.response;
+package webserver.http.response;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
